10 - Refactor with Typescript reference
