+++ understanding-typescript/10-modules-namespaces/src/components/project-input.ts
@@ -1,0 +1,84 @@
+/// <reference path="base-component.ts" />
+/// <reference path="../decorators/autobind.ts" />
+/// <reference path="../util/validation.ts" />
+/// <reference path="../state/project-state.ts" />
+
+namespace App {
+  export class ProjectInput extends Component<HTMLDivElement, HTMLElement> {
+    private titleInputElement: HTMLInputElement;
+    private descriptionInputElement: HTMLInputElement;
+    private peopleInputElement: HTMLInputElement;
+
+    constructor(templateId: string, hostId: string) {
+      super(templateId, hostId, true, 'user-input');
+
+      this.titleInputElement = this.element.querySelector('#title')! as HTMLInputElement;
+      this.descriptionInputElement = this.element.querySelector('#description')! as HTMLInputElement;
+      this.peopleInputElement = this.element.querySelector('#people')! as HTMLInputElement;
+
+      this.configure();
+    }
+
+    configure() {
+      this.element.addEventListener('submit', this.submitHandler)
+    }
+
+    renderContent() {};
+
+    private getherUserInput(): [title: string, description: string, people: number] | void {
+      const enteredTitle = this.titleInputElement.value;
+      const enteredDescription = this.descriptionInputElement.value;
+      const enteredPeople = this.peopleInputElement.value;
+
+      const titleValidatable: Validatable = {
+        value: enteredTitle,
+        require: true
+      }
+
+      const desctiptionValidatable: Validatable = {
+        value: enteredDescription,
+        require: true,
+        minLength: 5
+      }
+
+      const peopleValidatable: Validatable = {
+        value: enteredPeople,
+        require: true,
+        min: 1
+      }
+
+      // if(enteredTitle.trim().length === 0 || enteredDescription.trim().length === 0 || enteredPeople.trim().length === 0){
+      if(
+          !validate(titleValidatable) ||
+          !validate(desctiptionValidatable) ||
+          !validate(peopleValidatable)
+        ) {
+        // alert('Error! Fields must have value')
+        console.error('Error! Fields must have value')
+        return;
+      }
+
+      return [enteredTitle, enteredDescription, +enteredPeople];
+    };
+
+    private clearInputs():void {
+      this.titleInputElement.value = '';
+      this.descriptionInputElement.value = '';
+      this.peopleInputElement.value = '';
+    }
+
+    @AutoBind
+    private submitHandler(event: Event) {
+      event.preventDefault();
+
+      
+      const userInput = this.getherUserInput();
+      if(Array.isArray(userInput)) {
+        const [title, description, people] = userInput;
+        projectState.addProject(title, description, people);
+        this.clearInputs();
+      }
+    }
+  }
+  
+}
