+++ understanding-typescript/10-modules-namespaces/src/components/project-item.ts
@@ -1,0 +1,45 @@
+/// <reference path="base-component.ts" />
+/// <reference path="../decorators/autobind.ts" />
+/// <reference path="../modules/project.ts" />
+/// <reference path="../modules/drag-drop.ts" />
+
+namespace App {
+  // ProjectItem Class 
+  export class ProjectItem extends Component<HTMLUListElement, HTMLLIElement> implements Draggable {
+    private project: Project;
+
+    get persons() {
+      return this.project.people === 1 ? '1 person' : `${this.project.people} persons`;
+    }
+
+    constructor(hostId: string, project: Project) {
+      super('single-project', hostId, false, project.id)
+      this.project = project;
+
+      this.configure();
+      this.renderContent();
+    }
+
+    renderContent() {
+      this.element.querySelector('h2')!.textContent = this.project.title;
+      this.element.querySelector('h3')!.textContent = this.persons + ' assigned';
+      this.element.querySelector('p')!.textContent = this.project.description;
+    }
+
+    configure() {
+      this.element.addEventListener('dragstart', this.dragStartHandler);
+      this.element.addEventListener('dragend', this.dragEndHandler);
+    }
+
+    @AutoBind
+    dragStartHandler(event: DragEvent): void {
+      event.dataTransfer!.setData('text/plain', this.project.id);
+      event.dataTransfer!.effectAllowed = 'move';
+    }
+
+    dragEndHandler(_: DragEvent): void {
+      console.log('DragEnd')
+    }
+  }
+  
+}
