+++ understanding-typescript/10-modules-namespaces/src/decorators/autobind.ts
@@ -1,0 +1,15 @@
+namespace App {
+  export function AutoBind(_: any, _1: string, descriptor: PropertyDescriptor) {
+    const originalMethod = descriptor.value;
+    const adjDescriptor = {
+      configurable: true,
+      get() {
+        const boundFn = originalMethod.bind(this);
+
+        return boundFn;
+      }
+    }
+
+    return adjDescriptor;  
+  }
+}
