+++ understanding-typescript/10-modules-namespaces/src/modules/drag-drop.ts
@@ -1,0 +1,12 @@
+namespace App {
+  export interface Draggable {
+    dragStartHandler(event: DragEvent): void;
+    dragEndHandler(event: DragEvent): void;
+  }
+  
+  export interface DragTarget {
+    dragOverHandler(event: DragEvent): void;
+    dropHandler(event: DragEvent): void;
+    dragLeaveHandler(event: DragEvent): void;
+  }
+}
