+++ understanding-typescript/10-modules-namespaces/src/modules/project.ts
@@ -1,0 +1,18 @@
+namespace App {
+  export enum ProjectStatus { Active, Finished }
+
+  // Project type
+  export class Project {
+    constructor(
+      public id: string,
+      public title: string, 
+      public description: string, 
+      public people: number, 
+      public status: ProjectStatus) {
+      this.id = Math.random().toString();
+      this.title = title;
+      this.description = description;
+      this.people = people;
+    }
+  }
+}
