+++ understanding-typescript/10-modules-namespaces/src/state/project-state.ts
@@ -1,0 +1,60 @@
+namespace App {
+  // Project State Managment 
+  type Listener<T> = (items: T[]) => void;
+
+  class State<T> {
+    protected listeners: Listener<T>[] = [];
+
+    addListener(listenerFn: Listener<T>) {
+      this.listeners.push(listenerFn);
+    }
+  }
+
+  export class ProjectState extends State<Project> {
+    projects: Project[] = [];
+    private static instance: ProjectState;
+
+    private constructor() {
+      super();
+    }
+
+    static getInstance() {
+      if(!this.instance) {
+        this.instance = new ProjectState();
+      } 
+      return this.instance;
+    }
+
+    addProject(title: string, description: string, people: number) {
+      const newProject = new Project(
+        Math.random().toString(),
+        title,
+        description,
+        people,
+        ProjectStatus.Active  
+      )
+
+      this.projects.push(newProject);
+
+      this.updateListeners();
+    }
+
+    moveProject(projectId: string, newStatus: ProjectStatus) {
+      const project = this.projects.find(el => el.id === projectId);
+
+      if(project && project.status !== newStatus) {
+        project.status = newStatus;
+        this.updateListeners();
+      }
+    }
+
+    private updateListeners() {    
+      for(const listenerFn of this.listeners) {
+        // Pass a copy ot the array by calling .slice()
+        listenerFn(this.projects.slice()) // To play around with referrence
+      }
+    }
+  }
+
+  export const projectState = ProjectState.getInstance();
+}
